Move GagGlobals to gags
