+++ battle/Battle.py
@@ -13,9 +13,9 @@ from core.Entity import BattleEntity, Entity
 from core.Exceptions import (CogAlreadyTrappedError, CogLuredError, Error,
                              InvalidCogAttackTarget, TooManyCogsError,
                              TooManyToonsError)
-from core.GagGlobals import TRACK
 from core.Toon import Toon
 from gags.Gag import Gag, get_gag_track_name
+from gags.GagGlobals import TRACK
 
 # TODO Look into Strategy design patterns for Toon decision making
 # TODO #38 Different strategies: max_reward, fast_win_ignore_reward, survive..
