+++ core/Toon.py
@@ -4,9 +4,9 @@ from dataclasses import dataclass, field
 from typing import Optional
 
 from core.Entity import Entity
-from core.GagGlobals import DEFAULT_GAG_LIMIT
 from core.ToonGlobals import DEFAULT_BEAN_COUNT, DEFAULT_BEAN_LIMIT, DEFAULT_HP
 from gags.Gag import Gags
+from gags.GagGlobals import DEFAULT_GAG_LIMIT
 
 
 @dataclass
