+++ gags/GagGlobals.py
