+++ tests/battle/test_battle_toon_get_attacks.py
@@ -3,15 +3,15 @@ import pytest
 from battle.Battle import Battle, BattleCog, BattleToon, RewardCalculator
 from core.Cog import Cog
 from core.Factory import BattleCogFactory, CogFactory
-from core.GagGlobals import GAG, TRACK
+from core.Toon import Toon
 from core.ToonGlobals import (ASTRO_EXPECTED_AVAILABLE_GAGS,
                               ASTRO_EXPECTED_TRACK_LEVELS,
                               ASTRO_EXPECTED_UNLOCKED_GAGS,
                               TRAPA_EXPECTED_AVAILABLE_GAGS,
                               TRAPA_EXPECTED_TRACK_LEVELS,
                               TRAPA_EXPECTED_UNLOCKED_GAGS)
+from gags.GagGlobals import GAG, TRACK
 from tests.fixtures.battle_fixtures import get_expected_reward
-from core.Toon import Toon
 
 BATTLE_ID = 1
